requirements.txt
beautifulsoup4==4.9.3
Pillow==8.3.1
